OC-531 ~ Fixes UNKNOWN_DEVICE exceptions
UNKNOWN_DEVICE exceptions were thrown in OSGP Core because the device
has not been added yet in the database, caused by storing the device and
sending a queue message within the same transaction in the domain
adapter. (It seems like the message was picked up by OSGP Core before
the transaction in the domain adapter was committed)
