+++ osgp/platform/osgp-adapter-domain-smartmetering/pom.xml
@@ -101,6 +101,12 @@
       <artifactId>spring-data-jpa</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-test</artifactId>
+      <scope>test</scope>
+    </dependency>
+
     <!-- Hikari connection pooling -->
     <dependency>
       <groupId>com.zaxxer</groupId>
