+++ osgp/platform/osgp-adapter-domain-smartmetering/src/main/java/org/opensmartgridplatform/adapter/domain/smartmetering/application/services/InstallationService.java
@@ -11,19 +11,7 @@ package org.opensmartgridplatform.adapter.domain.smartmetering.application.servi
 import org.opensmartgridplatform.adapter.domain.smartmetering.application.mapping.CommonMapper;
 import org.opensmartgridplatform.adapter.domain.smartmetering.infra.jms.core.OsgpCoreRequestMessageSender;
 import org.opensmartgridplatform.adapter.domain.smartmetering.infra.jms.ws.WebServiceResponseMessageSender;
-import org.opensmartgridplatform.domain.core.entities.DeviceAuthorization;
-import org.opensmartgridplatform.domain.core.entities.DeviceModel;
-import org.opensmartgridplatform.domain.core.entities.Manufacturer;
-import org.opensmartgridplatform.domain.core.entities.Organisation;
-import org.opensmartgridplatform.domain.core.entities.ProtocolInfo;
 import org.opensmartgridplatform.domain.core.entities.SmartMeter;
-import org.opensmartgridplatform.domain.core.repositories.DeviceAuthorizationRepository;
-import org.opensmartgridplatform.domain.core.repositories.DeviceModelRepository;
-import org.opensmartgridplatform.domain.core.repositories.ManufacturerRepository;
-import org.opensmartgridplatform.domain.core.repositories.OrganisationRepository;
-import org.opensmartgridplatform.domain.core.repositories.ProtocolInfoRepository;
-import org.opensmartgridplatform.domain.core.repositories.SmartMeterRepository;
-import org.opensmartgridplatform.domain.core.valueobjects.DeviceFunctionGroup;
 import org.opensmartgridplatform.domain.core.valueobjects.smartmetering.AddSmartMeterRequest;
 import org.opensmartgridplatform.domain.core.valueobjects.smartmetering.CoupleMbusDeviceByChannelRequestData;
 import org.opensmartgridplatform.domain.core.valueobjects.smartmetering.CoupleMbusDeviceByChannelResponse;
@@ -34,9 +22,7 @@ import org.opensmartgridplatform.dto.valueobjects.smartmetering.CoupleMbusDevice
 import org.opensmartgridplatform.dto.valueobjects.smartmetering.DeCoupleMbusDeviceResponseDto;
 import org.opensmartgridplatform.dto.valueobjects.smartmetering.MbusChannelElementsResponseDto;
 import org.opensmartgridplatform.dto.valueobjects.smartmetering.SmartMeteringDeviceDto;
-import org.opensmartgridplatform.shared.exceptionhandling.ComponentType;
 import org.opensmartgridplatform.shared.exceptionhandling.FunctionalException;
-import org.opensmartgridplatform.shared.exceptionhandling.FunctionalExceptionType;
 import org.opensmartgridplatform.shared.exceptionhandling.OsgpException;
 import org.opensmartgridplatform.shared.infra.jms.DeviceMessageMetadata;
 import org.opensmartgridplatform.shared.infra.jms.RequestMessage;
@@ -52,7 +38,6 @@ import org.springframework.transaction.annotation.Transactional;
 import ma.glasnost.orika.MapperFactory;
 
 @Service(value = "domainSmartMeteringInstallationService")
-@Transactional(value = "transactionManager")
 public class InstallationService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(InstallationService.class);
@@ -62,18 +47,6 @@ public class InstallationService {
     private OsgpCoreRequestMessageSender osgpCoreRequestMessageSender;
 
     @Autowired
-    private SmartMeterRepository smartMeteringDeviceRepository;
-
-    @Autowired
-    private ManufacturerRepository manufacturerRepository;
-
-    @Autowired
-    private DeviceModelRepository deviceModelRepository;
-
-    @Autowired
-    private ProtocolInfoRepository protocolInfoRepository;
-
-    @Autowired
     private MapperFactory mapperFactory;
 
     @Autowired
@@ -81,10 +54,8 @@ public class InstallationService {
     private WebServiceResponseMessageSender webServiceResponseMessageSender;
 
     @Autowired
-    private OrganisationRepository organisationRepository;
-
-    @Autowired
-    private DeviceAuthorizationRepository deviceAuthorizationRepository;
+    @Qualifier(value = "domainSmartMeteringSmartMeterService")
+    private SmartMeterService smartMeterService;
 
     @Autowired
     private MBusGatewayService mBusGatewayService;
@@ -103,51 +74,26 @@ public class InstallationService {
         LOGGER.debug("addMeter for organisationIdentification: {} for deviceIdentification: {}", organisationId,
                 deviceId);
         final SmartMeteringDevice smartMeteringDevice = addSmartMeterRequest.getDevice();
-        final SmartMeter smartMeter = this.getSmartMeter(deviceId, smartMeteringDevice);
-        this.addMeter(organisationId, addSmartMeterRequest, smartMeter);
+
+        final SmartMeter smartMeter = this.smartMeterService.getSmartMeter(deviceId, smartMeteringDevice);
+        this.smartMeterService.storeMeter(organisationId, addSmartMeterRequest, smartMeter);
         this.osgpCoreRequestMessageSender.send(this.getRequestMessage(deviceMessageMetadata, smartMeteringDevice),
                 deviceMessageMetadata.getMessageType(), deviceMessageMetadata.getMessagePriority(),
                 deviceMessageMetadata.getScheduleTime());
     }
 
-    private SmartMeter getSmartMeter(final String deviceId, final SmartMeteringDevice smartMeteringDevice)
-            throws FunctionalException {
-        if (this.smartMeteringDeviceRepository.findByDeviceIdentification(deviceId) != null) {
-            throw new FunctionalException(FunctionalExceptionType.EXISTING_DEVICE, ComponentType.DOMAIN_SMART_METERING);
-        }
-        return this.mapperFactory.getMapperFacade().map(smartMeteringDevice, SmartMeter.class);
-    }
+    /**
+     * In case of errors that prevented adding the meter to the protocol
+     * database, the meter should be removed from the core database as well.
+     */
+    public void removeMeter(final DeviceMessageMetadata deviceMessageMetadata) {
+        LOGGER.warn(
+                "Removing meter {} for organization {}, because adding it to the protocol database failed with "
+                        + "correlation UID {}",
+                deviceMessageMetadata.getDeviceIdentification(), deviceMessageMetadata.getOrganisationIdentification(),
+                deviceMessageMetadata.getCorrelationUid());
 
-    private void addMeter(final String organisationIdentification, final AddSmartMeterRequest addSmartMeterRequest,
-            SmartMeter smartMeter) throws FunctionalException {
-        final SmartMeteringDevice smartMeteringDevice = addSmartMeterRequest.getDevice();
-        smartMeter.updateProtocol(this.getProtocolInfo(smartMeteringDevice));
-        smartMeter.setDeviceModel(this.getDeviceModel(addSmartMeterRequest.getDeviceModel()));
-        smartMeter = this.smartMeteringDeviceRepository.save(smartMeter);
-        this.storeAuthorization(organisationIdentification, smartMeter);
-    }
-
-    private ProtocolInfo getProtocolInfo(final SmartMeteringDevice smartMeteringDevice) throws FunctionalException {
-        final ProtocolInfo protocolInfo = this.protocolInfoRepository.findByProtocolAndProtocolVersion(
-                smartMeteringDevice.getProtocolName(), smartMeteringDevice.getProtocolVersion());
-        if (protocolInfo == null) {
-            throw new FunctionalException(FunctionalExceptionType.UNKNOWN_PROTOCOL_NAME_OR_VERSION,
-                    ComponentType.DOMAIN_SMART_METERING);
-        }
-        return protocolInfo;
-    }
-
-    private DeviceModel
-            getDeviceModel(final org.opensmartgridplatform.domain.core.valueobjects.DeviceModel deviceModel) {
-        final Manufacturer manufacturer = this.manufacturerRepository.findByCode(deviceModel.getManufacturer());
-        return this.deviceModelRepository.findByManufacturerAndModelCode(manufacturer, deviceModel.getModelCode());
-    }
-
-    private void storeAuthorization(final String organisationIdentification, final SmartMeter smartMeter) {
-        final Organisation organisation = this.organisationRepository
-                .findByOrganisationIdentification(organisationIdentification);
-        final DeviceAuthorization authorization = smartMeter.addAuthorization(organisation, DeviceFunctionGroup.OWNER);
-        this.deviceAuthorizationRepository.save(authorization);
+        this.smartMeterService.removeMeter(deviceMessageMetadata);
     }
 
     private RequestMessage getRequestMessage(final DeviceMessageMetadata deviceMessageMetadata,
@@ -157,26 +103,7 @@ public class InstallationService {
                 this.mapperFactory.getMapperFacade().map(smartMeteringDevice, SmartMeteringDeviceDto.class));
     }
 
-    /**
-     * In case of errors that prevented adding the meter to the protocol
-     * database, the meter should be removed from the core database as well.
-     */
-    @Transactional
-    public void removeMeter(final DeviceMessageMetadata deviceMessageMetadata) {
-
-        final SmartMeter device = this.smartMeteringDeviceRepository
-                .findByDeviceIdentification(deviceMessageMetadata.getDeviceIdentification());
-
-        LOGGER.warn(
-                "Removing meter {} for organization {}, because adding it to the protocol database failed with "
-                        + "correlation UID {}",
-                deviceMessageMetadata.getDeviceIdentification(), deviceMessageMetadata.getOrganisationIdentification(),
-                deviceMessageMetadata.getCorrelationUid());
-
-        this.deviceAuthorizationRepository.delete(device.getAuthorizations());
-        this.smartMeteringDeviceRepository.delete(device);
-    }
-
+    @Transactional(value = "transactionManager")
     public void handleAddMeterResponse(final DeviceMessageMetadata deviceMessageMetadata,
             final ResponseMessageResultType deviceResult, final OsgpException exception) {
         this.handleResponse("handleDefaultDeviceResponse", deviceMessageMetadata, deviceResult, exception);
@@ -187,16 +114,19 @@ public class InstallationService {
         this.mBusGatewayService.coupleMbusDevice(deviceMessageMetadata, requestData);
     }
 
+    @Transactional(value = "transactionManager")
     public void deCoupleMbusDevice(final DeviceMessageMetadata deviceMessageMetadata,
             final DeCoupleMbusDeviceRequestData requestData) throws FunctionalException {
         this.mBusGatewayService.deCoupleMbusDevice(deviceMessageMetadata, requestData);
     }
 
+    @Transactional(value = "transactionManager")
     public void coupleMbusDeviceByChannel(final DeviceMessageMetadata deviceMessageMetadata,
             final CoupleMbusDeviceByChannelRequestData requestData) throws FunctionalException {
         this.mBusGatewayService.coupleMbusDeviceByChannel(deviceMessageMetadata, requestData);
     }
 
+    @Transactional(value = "transactionManager")
     public void handleCoupleMbusDeviceResponse(final DeviceMessageMetadata deviceMessageMetadata,
             final ResponseMessageResultType result, final OsgpException exception,
             final MbusChannelElementsResponseDto dataObject) throws FunctionalException {
@@ -206,6 +136,7 @@ public class InstallationService {
         this.handleResponse("coupleMbusDevice", deviceMessageMetadata, result, exception);
     }
 
+    @Transactional(value = "transactionManager")
     public void handleDeCoupleMbusDeviceResponse(final DeviceMessageMetadata deviceMessageMetadata,
             final ResponseMessageResultType result, final OsgpException exception,
             final DeCoupleMbusDeviceResponseDto deCoupleMbusDeviceResponseDto) throws FunctionalException {
@@ -215,6 +146,7 @@ public class InstallationService {
         this.handleResponse("deCoupleMbusDevice", deviceMessageMetadata, result, exception);
     }
 
+    @Transactional(value = "transactionManager")
     public void handleCoupleMbusDeviceByChannelResponse(final DeviceMessageMetadata deviceMessageMetadata,
             final ResponseMessageResultType responseMessageResultType, final OsgpException osgpException,
             final CoupleMbusDeviceByChannelResponseDto dataObject) throws FunctionalException {
@@ -234,6 +166,7 @@ public class InstallationService {
         this.webServiceResponseMessageSender.send(responseMessage, deviceMessageMetadata.getMessageType());
     }
 
+    @Transactional(value = "transactionManager")
     public void handleResponse(final String methodName, final DeviceMessageMetadata deviceMessageMetadata,
             final ResponseMessageResultType deviceResult, final OsgpException exception) {
